+++ sdk/azurestackhci/arm-azurestackhci/src/operationsInterfaces/arcSettings.ts
@@ -15,7 +15,11 @@ import {
   ArcSettingsGetResponse,
   ArcSettingsCreateOptionalParams,
   ArcSettingsCreateResponse,
-  ArcSettingsDeleteOptionalParams
+  ArcSettingsDeleteOptionalParams,
+  ArcSettingsGeneratePasswordOptionalParams,
+  ArcSettingsGeneratePasswordResponse,
+  ArcSettingsCreateIdentityOptionalParams,
+  ArcSettingsCreateIdentityResponse
 } from "../models";
 
 /// <reference lib="esnext.asynciterable" />
@@ -86,4 +90,48 @@ export interface ArcSettings {
     arcSettingName: string,
     options?: ArcSettingsDeleteOptionalParams
   ): Promise<void>;
+  /**
+   * Generate password for arc settings.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param clusterName The name of the cluster.
+   * @param arcSettingName The name of the proxy resource holding details of HCI ArcSetting information.
+   * @param options The options parameters.
+   */
+  generatePassword(
+    resourceGroupName: string,
+    clusterName: string,
+    arcSettingName: string,
+    options?: ArcSettingsGeneratePasswordOptionalParams
+  ): Promise<ArcSettingsGeneratePasswordResponse>;
+  /**
+   * Create Aad identity for arc settings.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param clusterName The name of the cluster.
+   * @param arcSettingName The name of the proxy resource holding details of HCI ArcSetting information.
+   * @param options The options parameters.
+   */
+  beginCreateIdentity(
+    resourceGroupName: string,
+    clusterName: string,
+    arcSettingName: string,
+    options?: ArcSettingsCreateIdentityOptionalParams
+  ): Promise<
+    PollerLike<
+      PollOperationState<ArcSettingsCreateIdentityResponse>,
+      ArcSettingsCreateIdentityResponse
+    >
+  >;
+  /**
+   * Create Aad identity for arc settings.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param clusterName The name of the cluster.
+   * @param arcSettingName The name of the proxy resource holding details of HCI ArcSetting information.
+   * @param options The options parameters.
+   */
+  beginCreateIdentityAndWait(
+    resourceGroupName: string,
+    clusterName: string,
+    arcSettingName: string,
+    options?: ArcSettingsCreateIdentityOptionalParams
+  ): Promise<ArcSettingsCreateIdentityResponse>;
 }